Refactor minnisblad routes: consolidate token validation and document length checks into utility functions
